LayoutTests/platform/mac-yosemite/fast/canvas/canvas-too-large-to-draw-expected.txt

CONSOLE MESSAGE: line 36: Canvas area exceeds the maximum limit (width * height > 67108864).
 
